+++ .eslintrc.js
@@ -1,6 +1,6 @@
 module.exports = {
     extends: ['eslint:recommended'], // extending recommended config and config derived from eslint-config-prettier
-    plugins: ['prettier'], // activating esling-plugin-prettier (--fix stuff)
+    plugins: ['prettier'], // activating eslint-plugin-prettier (--fix stuff)
     parser: 'babel-eslint',
     globals: {
         WaveSurfer: true,
@@ -27,6 +27,7 @@ module.exports = {
             }
         ],
         eqeqeq: 'off',
+        'semi': 2,
         'no-console': 1,
         'no-unused-vars': 'off',
         'no-unreachable': 2,