+++ example/regions/app.js
@@ -17,11 +17,13 @@ document.addEventListener('DOMContentLoaded', function() {
                     {
                         start: 1,
                         end: 3,
+                        loop: false,
                         color: 'hsla(400, 100%, 30%, 0.5)'
                     },
                     {
                         start: 5,
                         end: 7,
+                        loop: false,
                         color: 'hsla(200, 50%, 70%, 0.4)'
                     }
                 ],
@@ -32,10 +34,14 @@ document.addEventListener('DOMContentLoaded', function() {
         ]
     });
 
+    wavesurfer.on('error', function(e) {
+        console.warn(e);
+    });
+
     // Load audio from URL
     wavesurfer.load('../media/demo.wav');
 
-    // this is already being done in /examples/trivia.js
+    // this is already being done in /example/trivia.js
     // document.querySelector(
     //     '[data-action="play"]'
     // ).addEventListener('click', wavesurfer.playPause.bind(wavesurfer));