+++ src/drawer.js
@@ -123,7 +123,7 @@ export default class Drawer extends util.Observer {
         this.wrapper.addEventListener('click', e => {
             const scrollbarHeight =
                 this.wrapper.offsetHeight - this.wrapper.clientHeight;
-            if (scrollbarHeight != 0) {
+            if (scrollbarHeight !== 0) {
                 // scrollbar is visible.  Check if click was on it
                 const bbox = this.wrapper.getBoundingClientRect();
                 if (e.clientY >= bbox.bottom - scrollbarHeight) {
@@ -134,6 +134,12 @@ export default class Drawer extends util.Observer {
 
             if (this.params.interact) {
                 this.fireEvent('click', e, this.handleEvent(e));
+            }
+        });
+
+        this.wrapper.addEventListener('dblclick', e => {
+            if (this.params.interact) {
+                this.fireEvent('dblclick', e, this.handleEvent(e));
             }
         });
 
@@ -203,8 +209,13 @@ export default class Drawer extends util.Observer {
 
         // if the cursor is currently visible...
         if (!immediate && -half <= offset && offset < half) {
-            // we'll limit the "re-center" rate.
-            const rate = 5;
+            // set rate at which waveform is centered
+            let rate = this.params.autoCenterRate;
+
+            // make rate depend on width of view and length of waveform
+            rate /= half;
+            rate *= maxScroll;
+
             offset = Math.max(-rate, Math.min(rate, offset));
             target = scrollLeft + offset;
         }
@@ -314,7 +325,10 @@ export default class Drawer extends util.Observer {
 
             if (this.params.scrollParent && this.params.autoCenter) {
                 const newPos = ~~(this.wrapper.scrollWidth * progress);
-                this.recenterOnPosition(newPos);
+                this.recenterOnPosition(
+                    newPos,
+                    this.params.autoCenterImmediately
+                );
             }
 
             this.updateProgress(pos);