+++ src/drawer.multicanvas.js
@@ -82,6 +82,14 @@ export default class MultiCanvas extends Drawer {
          * @type {number}
          */
         this.overlap = 2 * Math.ceil(params.pixelRatio / 2);
+
+        /**
+         * The radius of the wave bars. Makes bars rounded
+         *
+         * @private
+         * @type {number}
+         */
+        this.barRadius = params.barRadius || 0;
     }
 
     /**
@@ -301,12 +309,19 @@ export default class MultiCanvas extends Drawer {
                 for (i; i < last; i += step) {
                     const peak =
                         peaks[Math.floor(i * scale * peakIndexScale)] || 0;
-                    const h = Math.round((peak / absmax) * halfH);
+                    let h = Math.round((peak / absmax) * halfH);
+
+                    /* in case of silences, allow the user to specify that we
+                     * always draw *something* (normally a 1px high bar) */
+                    if (h == 0 && this.params.barMinHeight)
+                        h = this.params.barMinHeight;
+
                     this.fillRect(
                         i + this.halfPixel,
                         halfH - h + offsetY,
                         bar + this.halfPixel,
-                        h * 2
+                        h * 2,
+                        this.barRadius
                     );
                 }
             }
@@ -355,7 +370,8 @@ export default class MultiCanvas extends Drawer {
                     0,
                     halfH + offsetY - this.halfPixel,
                     this.width,
-                    this.halfPixel
+                    this.halfPixel,
+                    this.barRadius
                 );
             }
         );
@@ -388,8 +404,9 @@ export default class MultiCanvas extends Drawer {
      * @param {number} y Y-position of the rectangle
      * @param {number} width Width of the rectangle
      * @param {number} height Height of the rectangle
+     * @param {number} radius Radius of the rectangle
      */
-    fillRect(x, y, width, height) {
+    fillRect(x, y, width, height, radius) {
         const startCanvas = Math.floor(x / this.maxCanvasWidth);
         const endCanvas = Math.min(
             Math.ceil((x + width) / this.maxCanvasWidth) + 1,
@@ -417,7 +434,8 @@ export default class MultiCanvas extends Drawer {
                     intersection.x1 - leftOffset,
                     intersection.y1,
                     intersection.x2 - intersection.x1,
-                    intersection.y2 - intersection.y1
+                    intersection.y2 - intersection.y1,
+                    radius
                 );
             }
         }