+++ src/plugin/cursor.js
@@ -90,13 +90,14 @@ export default class CursorPlugin {
         const bbox = this.wavesurfer.container.getBoundingClientRect();
         let y = 0;
         let x = e.clientX - bbox.left;
+        let flip = bbox.right < e.clientX + this.outerWidth(this.displayTime);
 
         if (this.params.showTime && this.params.followCursorY) {
             // follow y-position of the mouse
             y = e.clientY - (bbox.top + bbox.height / 2);
         }
 
-        this.updateCursorPosition(x, y);
+        this.updateCursorPosition(x, y, flip);
     };
 
     /**
@@ -132,7 +133,7 @@ export default class CursorPlugin {
         /**
          * displays the time next to the cursor
          *
-         * @type {Boolean}
+         * @type {?HTMLElement}
          */
         this.showTime = null;
         /**
@@ -200,16 +201,21 @@ export default class CursorPlugin {
                         {
                             display: 'inline',
                             pointerEvents: 'none',
-                            margin: 'auto'
+                            margin: 'auto',
+                            visibility: 'hidden' // initial value will be hidden just for measuring purpose
                         },
                         this.params.customShowTimeStyle
                     )
                 )
             );
+            // initial value to measure display width
+            this.displayTime.innerHTML = this.formatTime(0);
         }
 
         this.wrapper.addEventListener('mousemove', this._onMousemove);
         if (this.params.hideOnBlur) {
+            // ensure elements are hidden initially
+            this.hideCursor();
             this.wrapper.addEventListener('mouseenter', this._onMouseenter);
             this.wrapper.addEventListener('mouseleave', this._onMouseleave);
         }
@@ -235,8 +241,9 @@ export default class CursorPlugin {
      *
      * @param {number} xpos The x offset of the cursor in pixels
      * @param {number} ypos The y offset of the cursor in pixels
+     * @param {boolean} flip Flag to flip duration text from right to left
      */
-    updateCursorPosition(xpos, ypos) {
+    updateCursorPosition(xpos, ypos, flip = false) {
         this.style(this.cursor, {
             left: `${xpos}px`
         });
@@ -253,9 +260,16 @@ export default class CursorPlugin {
             const timeValue =
                 Math.max(0, (xpos / elementWidth) * duration) + scrollTime;
             const formatValue = this.formatTime(timeValue);
+            if (flip) {
+                const textOffset = this.outerWidth(this.displayTime);
+                xpos -= textOffset;
+            }
             this.style(this.showTime, {
                 left: `${xpos}px`,
                 top: `${ypos}px`
+            });
+            this.style(this.displayTime, {
+                visibility: 'visible'
             });
             this.displayTime.innerHTML = `${formatValue}`;
         }
@@ -305,8 +319,24 @@ export default class CursorPlugin {
             [
                 Math.floor((time % 3600) / 60), // minutes
                 ('00' + Math.floor(time % 60)).slice(-2), // seconds
-                ('000' + Math.floor((time % 1) * 1000)).slice(-3) // miliseconds
+                ('000' + Math.floor((time % 1) * 1000)).slice(-3) // milliseconds
             ].join(':')
         );
+    }
+
+    /**
+     * Get outer width of given element.
+     *
+     * @param {DOM} element DOM Element
+     * @returns {number} outer width
+     */
+    outerWidth(element) {
+        if (!element) return 0;
+
+        let width = element.offsetWidth;
+        let style = getComputedStyle(element);
+
+        width += parseInt(style.marginLeft + style.marginRight);
+        return width;
     }
 }