+++ src/plugin/regions.js
@@ -24,6 +24,9 @@ class Region {
         this.resize =
             params.resize === undefined ? true : Boolean(params.resize);
         this.drag = params.drag === undefined ? true : Boolean(params.drag);
+        // reflect resize and drag state of region for region-updated listener
+        this.isResizing = false;
+        this.isDragging = false;
         this.loop = Boolean(params.loop);
         this.color = params.color || 'rgba(0, 0, 0, 0.1)';
         this.data = params.data || {};
@@ -94,16 +97,23 @@ class Region {
         }
     }
 
-    /* Play the audio region. */
-    play() {
-        this.wavesurfer.play(this.start, this.end);
+    /**
+     * Play the audio region.
+     * @param {number} start Optional offset to start playing at
+     */
+    play(start) {
+        const s = start || this.start;
+        this.wavesurfer.play(s, this.end);
         this.fireEvent('play');
         this.wavesurfer.fireEvent('region-play', this);
     }
 
-    /* Play the region in loop. */
-    playLoop() {
-        this.play();
+    /**
+     * Play the audio region in a loop.
+     * @param {number} start Optional offset to start playing at
+     * */
+    playLoop(start) {
+        this.play(start);
         this.once('out', () => this.playLoop());
     }
 
@@ -121,7 +131,6 @@ class Region {
             );
         }
 
-        const width = this.wrapper.scrollWidth;
         this.style(regionEl, {
             position: 'absolute',
             zIndex: 2,
@@ -142,16 +151,18 @@ class Region {
             const css = {
                 cursor: 'col-resize',
                 position: 'absolute',
-                left: '0px',
                 top: '0px',
                 width: '1%',
                 maxWidth: '4px',
                 height: '100%'
             };
             this.style(handleLeft, css);
+            this.style(handleLeft, {
+                left: '0px'
+            });
             this.style(handleRight, css);
             this.style(handleRight, {
-                left: '100%'
+                right: '0px'
             });
         }
 
@@ -230,18 +241,21 @@ class Region {
         this.firedOut = false;
 
         const onProcess = time => {
+            let start = Math.round(this.start * 10) / 10;
+            let end = Math.round(this.end * 10) / 10;
+            time = Math.round(time * 10) / 10;
+
             if (
                 !this.firedOut &&
                 this.firedIn &&
-                (this.start >= Math.round(time * 100) / 100 ||
-                    this.end <= Math.round(time * 100) / 100)
+                (start > time || end <= time)
             ) {
                 this.firedOut = true;
                 this.firedIn = false;
                 this.fireEvent('out');
                 this.wavesurfer.fireEvent('region-out', this);
             }
-            if (!this.firedIn && this.start <= time && this.end > time) {
+            if (!this.firedIn && start <= time && end > time) {
                 this.firedIn = true;
                 this.firedOut = false;
                 this.fireEvent('in');
@@ -292,7 +306,6 @@ class Region {
         (this.drag || this.resize) &&
             (() => {
                 const container = this.wavesurfer.drawer.container;
-                const duration = this.wavesurfer.getDuration();
                 const scrollSpeed = this.scrollSpeed;
                 const scrollThreshold = this.scrollThreshold;
                 let startTime;
@@ -306,6 +319,7 @@ class Region {
 
                 // Scroll when the user is dragging within the threshold
                 const edgeScroll = e => {
+                    const duration = this.wavesurfer.getDuration();
                     if (!scrollDirection || (!drag && !resize)) {
                         return;
                     }
@@ -335,6 +349,7 @@ class Region {
                 };
 
                 const onDown = e => {
+                    const duration = this.wavesurfer.getDuration();
                     if (e.touches && e.touches.length > 1) {
                         return;
                     }
@@ -358,7 +373,10 @@ class Region {
                         this.wrapper.scrollWidth - this.wrapper.clientWidth;
                     wrapperRect = this.wrapper.getBoundingClientRect();
 
-                    if (e.target.tagName.toLowerCase() == 'handle') {
+                    this.isResizing = false;
+                    this.isDragging = false;
+                    if (e.target.tagName.toLowerCase() === 'handle') {
+                        this.isResizing = true;
                         if (
                             e.target.classList.contains(
                                 'wavesurfer-handle-start'
@@ -369,6 +387,7 @@ class Region {
                             resize = 'end';
                         }
                     } else {
+                        this.isDragging = true;
                         drag = true;
                         resize = false;
                     }
@@ -379,6 +398,8 @@ class Region {
                     }
 
                     if (drag || resize) {
+                        this.isDragging = false;
+                        this.isResizing = false;
                         drag = false;
                         scrollDirection = null;
                         resize = false;
@@ -392,6 +413,8 @@ class Region {
                     }
                 };
                 const onMove = e => {
+                    const duration = this.wavesurfer.getDuration();
+
                     if (e.touches && e.touches.length > 1) {
                         return;
                     }
@@ -511,7 +534,7 @@ class Region {
     }
 
     onResize(delta, direction) {
-        if (direction == 'start') {
+        if (direction === 'start') {
             this.update({
                 start: Math.min(this.start + delta, this.end),
                 end: Math.max(this.start + delta, this.end)
@@ -535,6 +558,7 @@ class Region {
  * @property {?number} snapToGridInterval Snap the regions to a grid of the specified multiples in seconds
  * @property {?number} snapToGridOffset Shift the snap-to-grid by the specified seconds. May also be negative.
  * @property {?boolean} deferInit Set to true to manually call
+ * @property {number[]} maxRegions Maximum number of regions that may be created by the user at one time.
  * `initPlugin('regions')`
  */
 
@@ -627,6 +651,7 @@ export default class RegionsPlugin {
         this.params = params;
         this.wavesurfer = ws;
         this.util = ws.util;
+        this.maxRegions = params.maxRegions;
         this.util.getRegionSnapToGridValue = value => {
             return this.getRegionSnapToGridValue(value, params);
         };
@@ -652,6 +677,7 @@ export default class RegionsPlugin {
         // Id-based hash of regions
         this.list = {};
         this._onReady = () => {
+            this.wrapper = this.wavesurfer.drawer.wrapper;
             if (this.params.dragSelection) {
                 this.enableDragSelection(this.params);
             }
@@ -680,12 +706,25 @@ export default class RegionsPlugin {
     }
 
     /**
+     * check to see if adding a new region would exceed maxRegions
+     * @return {boolean} whether we should proceed and create a region
+     * @private
+     */
+    wouldExceedMaxRegions() {
+        return (
+            this.maxRegions && Object.keys(this.list).length >= this.maxRegions
+        );
+    }
+
+    /**
      * Add a region
      *
      * @param {object} params Region parameters
      * @return {Region} The created region
      */
     add(params) {
+        if (this.wouldExceedMaxRegions()) return null;
+
         const region = new this.wavesurfer.Region(params, this.wavesurfer);
 
         this.list[region.id] = region;
@@ -819,8 +858,10 @@ export default class RegionsPlugin {
                 return;
             }
 
+            // auto-create a region during mouse drag, unless region-count would exceed "maxRegions"
             if (!region) {
                 region = this.add(params || {});
+                if (!region) return;
             }
 
             const end = this.wavesurfer.drawer.handleEvent(e);