+++ CHANGES.md
@@ -1,11 +1,67 @@
 wavesurfer.js changelog
 =======================
 
-3.0.0 (unreleased)
+3.3.0 (unreleased)
+------------------
+
+- Add `wavesurfer.setPlayEnd(position)` to set a point in seconds for
+  playback to stop at (#1795)
+- Add `barMinHeight` option (#1693)
+- Expose progress to the `dblclick` event (#1790)
+- Regions plugin:
+  - Add `start` argument to `play` and `playLoop` methods (#1794)
+  - Add `maxRegions` option to limit max numbers of created regions (#1793)
+- Cursor plugin: fix time visibility (#1802)
+
+3.2.0 (24.10.2019)
+------------------
+
+- New `MediaElementWebAudio` backend (#1767):
+  - Allows you to use Web Audio API with big audio files, loading audio
+    like with MediaElement backend (HTML5 audio tag), so you can use the
+    same methods of MediaElement backend for loading and playback. This way,
+    the audio resource is not loaded entirely from server, but in ranges,
+    allowing you to use WebAudio features, like filters, on audio files with
+    a long duration. You can also supply peaks data, so the entire audio file
+    does not have to be decoded.
+    For example:
+    ```
+    wavesurfer.load(url | HTMLMediaElement, peaks, preload, duration);
+    wavesurfer.play();
+    wavesurfer.setFilter(customFilter);
+    ```
+- Add `barRadius` option to create waveforms with rounded bars (#953)
+- Throw error when the url parameter supplied to `wavesurfer.load()`
+  is empty (#1773, #1775)
+- Specify non-minified wavesurfer.js in `main` entry of `package.json` (#1759)
+- Add `dblclick` event listener to wavesurfer wrapper (#1764)
+- Fix `destroy()` in `MediaElement` backend (#1778)
+- Cursor plugin: flip position of time text to left of the cursor where needed
+  to improve readability (#1776)
+- Regions plugin: change region end handler position (#1762, #1781)
+
+3.1.0 (26.09.2019)
+------------------
+
+- Add `autoCenter` and `autoCenterRate` options (#1699)
+- Make sure `isReady` is true before firing the `ready` event (#1749)
+- Improve fetch error messages (#1748)
+- Use `MediaElement` backend for browsers that don't support WebAudio (#1739)
+- Regions plugin:
+  - Use `isResizing` and `isDragging` to filter events in
+    region-updated listener (#1716)
+  - Fix `playLoop` and `loop` option for clips with duration <15s (#1626)
+- Spectrogram plugin: fix variable name in click handler (#1742)
+- Minimap plugin: fix left/width calculations for regions on retina/4k
+  screens (#1743)
+- New example: video-annotation (#1726)
+
+3.0.0 (11.07.2019)
 ------------------
 
 - Add `wavesurfer.getActivePlugins()`: return map of plugins
   that are currently initialised
+- Replace usage of `util.ajax` with `util.fetchFile` (#1365)
 - Update progress when seeking with HTML media controls (#1535)
 - Make sure mute/volume is updated when using `MediaElement` backend (#1615)
 - Refactor `MultiCanvas` and add `CanvasEntry` class (#1617)
@@ -14,12 +70,14 @@ wavesurfer.js changelog
 - Add support for `Blob` output type in `wavesurfer.exportImage` (#1610)
 - Fix fallback to Audio Element in browsers that don't support Web Audio (#1614)
 - `util.getId()` now accepts a `prefix` argument (#1619)
+- Improve documentation for `xhr` option (#1656)
 - Fix: the `progressWave` should not be rendered when specifying the same
   value for the `progressColor` and `waveColor` options (#1620)
 - Cursor plugin:
   - Add `formatTimeCallback` option
   - Add `followCursorY` option (#1605)
   - Remove deprecated `enableCursor` method (#1646)
+  - Hide the cursor elements before first mouseover if `hideOnBlur` is set (#1663)
 - Spectrogram plugin:
   - Fix `ready` listener when loading multiple audio files (#1572)
   - Allow user to specify a colorMap (#1436)
@@ -28,10 +86,14 @@ wavesurfer.js changelog
   - Add `snapToGridInterval` and `snapToGridOffset` options (#1632)
   - Allow drawing regions over existing regions, if the underlying ones are not
     draggable or resizable (#1633)
+  - Calculate the duration at event time to allow predefined regions to be
+    dragged and resized (#1673)
   - Remove deprecated `initRegions` method (#1646)
 - Timeline plugin: fix `ready` listener when loading multiple
   audio files
 - Minimap plugin: remove deprecated `initMinimap` method (#1646)
+
+Check `UPGRADE.md` for backward incompatible changes since v2.x.
 
 2.2.1 (18.03.2019)
 ------------------